finished notifications and fixed all the bugs in the code

+++ database/migrations/2025_12_10_073058_create_ratings_table.php
@@ -20,7 +20,7 @@ return new class extends Migration
             $table->foreignId('property_id')->constrained()->cascadeOnDelete();
 
             // Rating details
-            $table->unsignedTinyInteger('stars'); // 1-5 stars
+            $table->decimal('stars',2,1); // 1-5 stars
             $table->string('comment')->nullable();
 
             // Ensure a user can rate a property only once
